pyproject.toml
[project]
name = "myragchatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Antonio Morabito",email = "antoniomorabitositorus@gmail.com"}
]
readme = "README.md"


[tool.poetry]
name = "myragchatbot"
version = "0.1.0"
description = "RAG chatbot with Chainlit, LangChain, Chroma"
authors = ["Antonio"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

# Core packages

langchain = "*"
langchain-google-genai ="2.1.2,<3.0.0"
cryptography =">=44.0.2,<45.0.0"
langchain-community = "*"
langchain-ollama = "*"
langchain-core = "^0.3"
langchain-openai = "*"
langchain-text-splitters = "*"
chromadb = ">=0.4.0,<0.7.0"
openai = "^1.14.3"
cohere = "^5.14.2"

# PDF/Docs
pypdf = "*"
unstructured = "*"

# Embedding
sentence-transformers = "^2.3.1"
huggingface-hub = "^0.22.2"

# Util
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
tqdm = "^4.66.2"

# UI
chainlit = "*"
streamlit = "1.31.1"
tavily-python = "^0.5.4"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
